with fragment_add_furniture.xml, fragment_edit_furniture.xml, fragment_floor_plan.xml, fragment_profile.xml

+++ app/src/main/java/postpc/finalproject/RoomInn/models/Chair.kt
@@ -1,7 +1,5 @@
 package postpc.finalproject.RoomInn.models
 
-import postpc.finalproject.RoomInn.Point3D
-
-class Chair(location : Point3D, length : Int = 40 , width : Int = 50, height : Int = 100,
-    color : String): Furniture(location,length,width, height,color) {
+class Chair(location : Point3D, length : Float = 40f, width : Float = 50f, height : Float = 100f,
+            color : String): Furniture(location,length,width, height,color) {
 }
