+++ app/src/main/java/postpc/finalproject/RoomInn/models/Closet.kt
@@ -1,7 +1,5 @@
 package postpc.finalproject.RoomInn.models
 
-import postpc.finalproject.RoomInn.Point3D
-
-class Closet(location : Point3D, length : Int = 100, width : Int = 50, height : Int = 200,
-              color : String): Furniture(location,length,width, height,color) {
+class Closet(location : Point3D, length : Float = 100f, width : Float = 50f, height : Float = 200f,
+             color : String): Furniture(location,length,width, height,color) {
 }
