+++ app/src/main/java/postpc/finalproject/RoomInn/models/Desk.kt
@@ -1,7 +1,5 @@
 package postpc.finalproject.RoomInn.models
 
-import postpc.finalproject.RoomInn.Point3D
-
-class Desk(location : Point3D, length : Int = 120, width : Int = 50, height : Int = 70,
+class Desk(location : Point3D, length : Float = 120f, width : Float = 50f, height : Float = 70f,
            color : String): Furniture(location,length,width, height,color) {
 }
