+++ app/src/main/java/postpc/finalproject/RoomInn/models/Window.kt
@@ -1,8 +1,6 @@
 package postpc.finalproject.RoomInn.models
 
-import postpc.finalproject.RoomInn.Point3D
-
 data class Window(var location : Point3D,
-                  var length : Int = 60,
-                  var width : Int = 80) {
+                  var length : Float = 60f,
+                  var width : Float = 80f) {
 }
